implement vault to record and get VCs
VCVault implemented and added to VCService to record and retrieve again VCs. The data is stored with JPA in a relational database.
The VCs are stored as JSON blob. Data is further stored in a normalized manner, most notably a dedicated claim table. This
allows to query the data (as long as it has a primitive type).